chore: clean up plugin code
